+++ src/main/java/com/github/hjgf0624/sideproject/entity/UserEntity.java
@@ -45,10 +45,10 @@ public class UserEntity {
 //    private String interests; // 관심사, TEXT
 
     @Column(name = "latitude", nullable = true)
-    private Float latitude = 0.0f; // 위도, FLOAT
+    private Double latitude = 0.0; // 위도, FLOAT
 
     @Column(name = "longitude", nullable = true)
-    private Float longitude = 0.0f; // 경도, FLOAT
+    private Double longitude = 0.0; // 경도, FLOAT
 
     @Column(name = "nickname", length = 60)
     private String nickname; // 닉네임, VARCHAR(60)
